WARNING: This commit *may* break. It will be followed by a series of commits that bring the repository into a working state.
This is a series of work-in-progress modifications to the legacy tokenizer that were being written when the decision was made to switch to LEPL for parsing. They are being committed here for the record, in case we need to revert back to this line of work. I cannot recall if these changes have been tested, but in any case it does not matter as they will be overwritten by the LEPL code in the very next commit.